[frontend] Some Root files updated from JS to TS (#7500)
