+++ opencti-platform/opencti-front/src/private/components/arsenal/malwares/MalwareKnowledge.jsx
@@ -11,7 +11,6 @@ import EntityStixSightingRelationships from '../../events/stix_sighting_relation
 import StixSightingRelationship from '../../events/stix_sighting_relationships/StixSightingRelationship';
 import EntityStixCoreRelationshipsIndicators from '../../common/stix_core_relationships/views/indicators/EntityStixCoreRelationshipsIndicators';
 import EntityStixCoreRelationshipsStixCyberObservable from '../../common/stix_core_relationships/views/stix_cyber_observable/EntityStixCoreRelationshipsStixCyberObservable';
-import withRouter from '../../../../utils/compat-router/withRouter';
 
 class MalwareKnowledgeComponent extends Component {
   render() {
@@ -297,4 +296,4 @@ const MalwareKnowledge = createFragmentContainer(MalwareKnowledgeComponent, {
   `,
 });
 
-export default withRouter(MalwareKnowledge);
+export default MalwareKnowledge;
