+++ opencti-platform/opencti-front/src/private/components/arsenal/vulnerabilities/VulnerabilityKnowledge.jsx
@@ -9,7 +9,6 @@ import EntityStixSightingRelationships from '../../events/stix_sighting_relation
 import StixSightingRelationship from '../../events/stix_sighting_relationships/StixSightingRelationship';
 import EntityStixCoreRelationshipsIndicators from '../../common/stix_core_relationships/views/indicators/EntityStixCoreRelationshipsIndicators';
 import EntityStixCoreRelationshipsStixCyberObservable from '../../common/stix_core_relationships/views/stix_cyber_observable/EntityStixCoreRelationshipsStixCyberObservable';
-import withRouter from '../../../../utils/compat-router/withRouter';
 
 class VulnerabilityKnowledgeComponent extends Component {
   render() {
@@ -240,4 +239,4 @@ const VulnerabilityKnowledge = createFragmentContainer(
   },
 );
 
-export default withRouter(VulnerabilityKnowledge);
+export default VulnerabilityKnowledge;
